Refactor image validation to use Zod schema for consistent client-side checks
