+++ components/form/FormContainer.tsx
@@ -4,7 +4,7 @@ import { actionFunction } from '@/utils/types'
 import { useRouter } from 'next/navigation'
 import React, { use, useActionState, useEffect, useTransition } from 'react'
 import { toast } from 'sonner'
-import { validateFileSize } from '@/utils/fileValidation'
+import { checkImageBeforeSubmit } from '@/utils/clientValidation'
 
 interface FormProps{
     children: React.ReactNode,
@@ -38,13 +38,12 @@ function FormContainer({children,action}:FormProps) {
     },[state])
 
     const handleSubmit = (formData: FormData) => {
-      // Check for file inputs and validate size before submission
+      // Check for file inputs and validate using existing schema before submission
       const imageFile = formData.get('image') as File;
-      if (imageFile && imageFile.size > 0) {
-        const { isValid } = validateFileSize(imageFile);
+      if (imageFile) {
+        const isValid = checkImageBeforeSubmit(imageFile);
         if (!isValid) {
-          toast.error('Please select a smaller image file before submitting');
-          return;
+          return; // Toast message is already shown in checkImageBeforeSubmit
         }
       }
